LPS-144606 Also move imports formatter and package classes to their own package, but within portal-impl
