+++ modules/util/portal-tools-java-parser/src/main/java/com/liferay/portal/tools/java/parser/JavaParser.java
@@ -29,9 +29,9 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.SystemProperties;
 import com.liferay.portal.kernel.util.Tuple;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.tools.ImportsFormatter;
-import com.liferay.portal.tools.JavaImportsFormatter;
 import com.liferay.portal.tools.ToolsUtil;
+import com.liferay.portal.tools.imports.ImportsFormatter;
+import com.liferay.portal.tools.imports.JavaImportsFormatter;
 import com.liferay.portal.tools.java.parser.util.DetailASTUtil;
 import com.liferay.portal.tools.java.parser.util.FileUtil;
 import com.liferay.portal.tools.java.parser.util.JavaParserUtil;
