+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/BNDIncludeResourceCheck.java
@@ -16,7 +16,7 @@ package com.liferay.source.formatter.checks;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.tools.ImportPackage;
+import com.liferay.portal.tools.imports.ImportPackage;
 import com.liferay.source.formatter.checks.util.BNDSourceUtil;
 
 import java.util.Comparator;
