+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/FTLImportsCheck.java
@@ -14,8 +14,8 @@
 
 package com.liferay.source.formatter.checks;
 
-import com.liferay.portal.tools.ImportsFormatter;
-import com.liferay.source.formatter.FTLImportsFormatter;
+import com.liferay.portal.tools.imports.FTLImportsFormatter;
+import com.liferay.portal.tools.imports.ImportsFormatter;
 
 import java.io.IOException;
