+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/checks/JSPImportsCheck.java
@@ -15,8 +15,8 @@
 package com.liferay.source.formatter.checks;
 
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.tools.ImportsFormatter;
-import com.liferay.source.formatter.JSPImportsFormatter;
+import com.liferay.portal.tools.imports.ImportsFormatter;
+import com.liferay.portal.tools.imports.JSPImportsFormatter;
 import com.liferay.source.formatter.checks.util.JSPSourceUtil;
 
 import java.io.IOException;
