+++ portal-impl/src/com/liferay/portal/tools/imports/BNDImportPackage.java
@@ -12,11 +12,10 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.portal.tools.imports;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
-import com.liferay.portal.tools.ImportPackage;
 
 /**
  * @author Alan Huang
