+++ portal-impl/src/com/liferay/portal/tools/imports/BNDImportsFormatter.java
@@ -12,13 +12,11 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.portal.tools.imports;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portal.tools.BaseImportsFormatter;
-import com.liferay.portal.tools.ImportPackage;
 
 import java.io.IOException;
