+++ portal-impl/src/com/liferay/portal/tools/imports/ImportsFormatter.java
@@ -12,33 +12,21 @@
  * details.
  */
 
-package com.liferay.source.formatter;
+package com.liferay.portal.tools.imports;
 
-import com.liferay.petra.string.StringPool;
-import com.liferay.portal.tools.ImportPackage;
+import java.io.IOException;
+
+import java.util.regex.Pattern;
 
 /**
- * @author Alan Huang
+ * @author Hugo Huijser
  */
-public class PythonImportPackage extends ImportPackage {
+public interface ImportsFormatter {
 
-	public PythonImportPackage(String importString, String line) {
-		super(importString, false, line);
+	public String format(String content, Pattern importPattern)
+		throws IOException;
 
-		_importString = importString;
-	}
-
-	@Override
-	public String getPackageLevel() {
-		int pos = _importString.indexOf(StringPool.PERIOD);
-
-		if (pos == -1) {
-			return _importString;
-		}
-
-		return _importString.substring(0, pos);
-	}
-
-	private final String _importString;
+	public String format(String content, String packageDir, String className)
+		throws IOException;
 
 }
