+++ portal-impl/src/com/liferay/portal/tools/imports/JSPImportsFormatter.java
@@ -1,0 +1,97 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.tools.imports;
+
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.Validator;
+
+import java.io.IOException;
+
+import java.util.ArrayList;
+import java.util.List;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+
+/**
+ * @author Carlos Sierra Andrés
+ * @author André de Oliveira
+ * @author Hugo Huijser
+ */
+public class JSPImportsFormatter extends BaseImportsFormatter {
+
+	public static List<String> getImportNames(String content) {
+		List<String> importNames = new ArrayList<>();
+
+		Matcher matcher = _jspImportPattern.matcher(content);
+
+		while (matcher.find()) {
+			importNames.add(matcher.group(1));
+		}
+
+		return importNames;
+	}
+
+	@Override
+	protected ImportPackage createImportPackage(String line) {
+		Matcher matcher = _jspImportPattern.matcher(line);
+
+		if (matcher.find()) {
+			return new ImportPackage(matcher.group(1), false, line);
+		}
+
+		matcher = _jspTaglibPattern.matcher(line);
+
+		if (matcher.find()) {
+			return new ImportPackage(matcher.group(1), false, line);
+		}
+
+		return null;
+	}
+
+	@Override
+	protected String doFormat(
+			String content, Pattern importPattern, String packageDir,
+			String className)
+		throws IOException {
+
+		String imports = getImports(content, importPattern);
+
+		if (Validator.isNull(imports)) {
+			return content;
+		}
+
+		String newImports = sortAndGroupImports(imports);
+
+		content = StringUtil.replaceFirst(content, imports, newImports + "\n");
+
+		return StringUtil.trimTrailing(content);
+	}
+
+	protected String getImports(String content, Pattern importPattern) {
+		Matcher matcher = importPattern.matcher(content);
+
+		if (matcher.find()) {
+			return matcher.group();
+		}
+
+		return null;
+	}
+
+	private static final Pattern _jspImportPattern = Pattern.compile(
+		"import=\"([^\\s\"]+)\"");
+	private static final Pattern _jspTaglibPattern = Pattern.compile(
+		"uri=\"http://([^\\s\"]+)\"");
+
+}
